requirements.txt
annotated-types==0.5.0
backcall==0.2.0
blis==0.7.11
cachetools==5.3.2
catalogue==1.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
cycler==0.11.0
cymem==2.0.8
daal==2023.1.1
daal4py==2023.1.1
debugpy==1.7.0
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
fastai==2.7.12
fastcore==1.5.29
fastdownload==0.0.7
fastprogress==1.0.3
fonttools==4.38.0
google-api-core==2.14.0
google-auth==2.25.1
google-cloud-core==2.3.3
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
idna==3.6
imageio==2.31.2
imbalanced-learn==0.11.0
importlib-metadata==6.7.0
ipykernel==6.16.2
ipython==7.34.0
jedi==0.19.1
joblib==1.3.2
jupyter_client==7.4.9
jupyter_core==4.12.0
kiwisolver==1.4.5
llvmlite==0.39.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
murmurhash==1.0.10
nest-asyncio==1.5.8
networkx==2.6.3
numba==0.56.4
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
openpyxl==3.1.2
packaging==23.2
pandas==1.3.5
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
plac==1.1.3
preshed==3.0.9
prompt-toolkit==3.0.41
protobuf==4.24.4
psutil==5.9.6
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pyts==0.12.0
pytz==2023.3.post1
PyWavelets==1.3.0
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scikit-learn-intelex==2023.1.1
scipy==1.7.3
six==1.16.0
spacy==2.3.9
srsly==1.0.7
TA-Lib==0.4.28
tbb==2021.11.0
thinc==7.4.6
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
tsai==0.3.4
typing_extensions==4.7.1
urllib3==2.0.7
wasabi==0.10.1
wcwidth==0.2.12
zipp==3.15.0
